Refactoring everything to be more modular
Taking an Entity-System approach instead of the traditional OOP approach
Everything is an Entity. There are three special methods that can be implemented:
input, update and render. If they are implemented in an Entity, they are going
to be called during the main game loop at the respective time.
Entities are being created in a very C-like manner, by creating the object and then
assigning attributes.